backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pandas==2.2.2
pydantic==2.8.2
python-multipart==0.0.9
boto3>=1.34.131
botocore==1.35.20
python-dotenv==1.0.1
mangum==0.17.0
openpyxl==3.1.5
